Feat: Image 16 -> 8 bit

+++ ScanSaveImg.m
@@ -1,13 +1,5 @@
-% 2021. 04. 01
-
-% easySCAN_v1.1.3 -> easySCAN_v1.1.4
-
-% Don't display time for scanning anymore.
-
-
 function ScanSaveImg(app, vid, FluorMode, scanCh, savDir, fileNum)
 
-% startScanSave = tic;
 chStr = {'BM' 'FM'};
 
 if length(scanCh) >= 2
@@ -16,14 +8,9 @@ if length(scanCh) >= 2
     im = Get_Snapshot(app, 1);   
     
     Ch_set(app, 2)
-    
-%     startSave = tic;
     savName = sprintf('%d_%s.png', fileNum, chStr{1});    
     pathFile = fullfile(savDir, savName);
     imwrite(im, pathFile)
-    
-%     timeSave = toc(startSave);
-%     fprintf('Time for Save = %d\n', timeSave) 
 
     im = Get_Snapshot(app, 2);   
 
@@ -43,16 +30,9 @@ if length(scanCh) >= 2
     im = im_colored;
 
     Ch_set(app, 1)
-%     startSave = tic;
     savName = sprintf('%d_%s.png', fileNum, chStr{2});    
     pathFile = fullfile(savDir, savName);
     imwrite(im, pathFile)
-    
-%     timeSave = toc(startSave);
-%     fprintf('Time for Save = %d\n', timeSave)    
-
-%     timeScanSave = toc(startScanSave);
-%     fprintf('Time for ScanSave = %d\n', timeScanSave)
     
 else
 
@@ -63,7 +43,8 @@ else
 
         if ch~=1
 
-            im_colored = uint16(zeros([size(im), 3]));
+%             im_colored = uint16(zeros([size(im), 3]));
+            im_colored = uint8(zeros([size(im), 3]));
 
             if strcmp(FluorMode, 'PE')
 
@@ -80,19 +61,9 @@ else
 
         end
 
-%         startSave = tic;
-
         savName = sprintf('%d_%s.png', fileNum, chStr{ch});    
         pathFile = fullfile(savDir, savName);
-
         imwrite(im, pathFile)
-
-%         timeSave = toc(startSave);
-%         fprintf('Time for Save = %d\n', timeSave)    
-
-%         timeScanSave = toc(startScanSave);
-%         fprintf('Time for ScanSave = %d\n', timeScanSave)
-
     end
     
 end
